# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/connects": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint creates a new connection by binding the incoming JSON request to the ConnectCreateInput model.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connects"
                ],
                "summary": "Create creates a new connection",
                "parameters": [
                    {
                        "description": "Connect creation input",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ConnectCreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Connect created successfully",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Connect creation failed",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            }
        },
        "/connects/{id}": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint updates a connection by binding the incoming JSON request to the ConnectUpdateInput model.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connects"
                ],
                "summary": "Update updates an existing connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Connection ID to update,approved:101, rejected:102",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Connect update input",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ConnectUpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Connect updated successfully",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Connect update failed",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.FailureResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controller.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "limit": {
                    "type": "integer"
                },
                "skip": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.ConnectCreateInput": {
            "type": "object",
            "required": [
                "friend_id"
            ],
            "properties": {
                "friend_id": {
                    "type": "string"
                }
            }
        },
        "model.ConnectUpdateInput": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "$ref": "#/definitions/model.RequestStatus"
                }
            }
        },
...
        "model.RequestStatus": {
            "type": "integer",
            "enum": [
                100,
                101,
                102
            ],
            "x-enum-varnames": [
                "RequestStatusPending",
                "RequestStatusApproved",
                "RequestStatusRejected"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \\\"Bearer \\\" and then your API Token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}